Rename audit-log to auditor
