+++ workspaces/rbac/plugins/rbac-backend/src/auditor/auditor.ts
