+++ workspaces/rbac/plugins/rbac-backend/src/auditor/rest-interceptor.ts
@@ -26,7 +26,7 @@ import {
   ListPluginPoliciesEvents,
   PermissionEvents,
   RoleEvents,
-} from './audit-logger';
+} from './auditor';
 import {
   AuditorService,
   AuditorServiceEvent,
